+++ sdk/spring/spring-cloud-azure-autoconfigure/src/main/java/com/azure/spring/cloud/autoconfigure/aad/configuration/AadOAuth2ClientConfiguration.java
@@ -31,10 +31,8 @@ import org.springframework.security.oauth2.client.OAuth2AuthorizedClientProvider
 import org.springframework.security.oauth2.client.RefreshTokenOAuth2AuthorizedClientProvider;
 import org.springframework.security.oauth2.client.endpoint.DefaultClientCredentialsTokenResponseClient;
 import org.springframework.security.oauth2.client.endpoint.DefaultJwtBearerTokenResponseClient;
-import org.springframework.security.oauth2.client.endpoint.DefaultPasswordTokenResponseClient;
 import org.springframework.security.oauth2.client.endpoint.DefaultRefreshTokenTokenResponseClient;
 import org.springframework.security.oauth2.client.endpoint.OAuth2ClientCredentialsGrantRequestEntityConverter;
-import org.springframework.security.oauth2.client.endpoint.OAuth2PasswordGrantRequestEntityConverter;
 import org.springframework.security.oauth2.client.endpoint.OAuth2RefreshTokenGrantRequestEntityConverter;
 import org.springframework.security.oauth2.client.registration.ClientRegistration;
 import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository;
@@ -134,7 +132,6 @@ public class AadOAuth2ClientConfiguration {
                         .authorizationCode()
                         .clientCredentials(builder ->
                                 clientCredentialsGrantBuilderAccessTokenResponseClientCustomizer(builder, jwkResolver))
-                        .password(builder -> passwordGrantBuilderAccessTokenResponseClientCustomizer(builder, jwkResolver))
                         .provider(refreshTokenProvider)
                         .provider(jwtBearerProvider)
                         .provider(azureDelegatedOAuth2AuthorizedClientProvider(refreshTokenProvider, authorizedClients))
@@ -192,19 +189,6 @@ public class AadOAuth2ClientConfiguration {
         AadAuthorizationServerEndpoints endpoints = new AadAuthorizationServerEndpoints(
                 profile.getEnvironment().getActiveDirectoryEndpoint(), profile.getTenantId());
         return new AadOidcIdTokenDecoderFactory(endpoints.getJwkSetEndpoint(), createRestTemplate(restTemplateBuilder));
-    }
-
-    private void passwordGrantBuilderAccessTokenResponseClientCustomizer(
-            OAuth2AuthorizedClientProviderBuilder.PasswordGrantBuilder builder,
-            OAuth2ClientAuthenticationJwkResolver resolver) {
-        DefaultPasswordTokenResponseClient client = new DefaultPasswordTokenResponseClient();
-        client.setRestOperations(createOAuth2AccessTokenResponseClientRestTemplate(restTemplateBuilder));
-        if (resolver != null) {
-            OAuth2PasswordGrantRequestEntityConverter converter = new OAuth2PasswordGrantRequestEntityConverter();
-            converter.addParametersConverter(new AadJwtClientAuthenticationParametersConverter<>(resolver::resolve));
-            client.setRequestEntityConverter(converter);
-        }
-        builder.accessTokenResponseClient(client);
     }
 
     private void clientCredentialsGrantBuilderAccessTokenResponseClientCustomizer(