+++ sdk/spring/spring-cloud-azure-autoconfigure/src/test/java/com/azure/spring/cloud/autoconfigure/aad/implementation/TestJwks.java
@@ -27,6 +27,7 @@ public final class TestJwks {
     private TestJwks() {
     }
 
+    @SuppressWarnings("deprecation")
     public static RSAKey.Builder jwk(RSAPublicKey publicKey, RSAPrivateKey privateKey, X509Certificate cert) {
         // @formatter:off
         try {