+++ sdk/spring/spring-cloud-azure-autoconfigure/src/test/java/com/azure/spring/cloud/autoconfigure/aad/implementation/webapi/AadResourceServerConfigurationTests.java
@@ -7,17 +7,19 @@ import com.azure.spring.cloud.autoconfigure.aad.properties.AadAuthenticationProp
 import com.nimbusds.jwt.proc.JWTClaimsSetAwareJWSKeySelector;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.FilteredClassLoader;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.oauth2.core.OAuth2TokenValidator;
 import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.security.oauth2.jwt.JwtDecoder;
 import org.springframework.security.oauth2.jwt.NimbusJwtDecoder;
 import org.springframework.security.oauth2.server.resource.BearerTokenAuthenticationToken;
+import org.springframework.security.oauth2.server.resource.web.BearerTokenAuthenticationFilter;
+import org.springframework.security.web.SecurityFilterChain;
 
 import java.util.List;
 
 import static com.azure.spring.cloud.autoconfigure.aad.implementation.WebApplicationContextRunnerUtils.resourceServerContextRunner;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 class AadResourceServerConfigurationTests {
@@ -70,13 +72,17 @@ class AadResourceServerConfigurationTests {
     }
 
     @Test
-    void testCreateWebSecurityConfigurerAdapter() {
+    void testResourceServerHttpSecurityConfigured() {
         resourceServerContextRunner()
             .withPropertyValues("spring.cloud.azure.active-directory.enabled=true")
             .run(context -> {
-                WebSecurityConfigurerAdapter webSecurityConfigurerAdapter = context
-                    .getBean(AadResourceServerConfiguration.DefaultAadResourceServerWebSecurityConfigurerAdapter.class);
-                assertThat(webSecurityConfigurerAdapter).isNotNull();
+                SecurityFilterChain filterChain = context.getBean(SecurityFilterChain.class);
+                assertThat(filterChain).isNotNull();
+                assertEquals(1,
+                    filterChain.getFilters()
+                               .stream()
+                               .filter(filter -> filter.getClass().equals(BearerTokenAuthenticationFilter.class))
+                               .count());
             });
     }
 }