+++ sdk/spring/spring-integration-azure-core/src/test/java/com/azure/spring/integration/core/handler/DefaultMessageHandlerTests.java
@@ -70,6 +70,7 @@ public abstract class DefaultMessageHandlerTests<O extends SendOperation> {
     }
 
     @Test
+    @SuppressWarnings("deprecation")
     public void testSendCallback() {
         ListenableFutureCallback<Void> callbackSpy = spy(new ListenableFutureCallback<Void>() {
             @Override